+++ backend/src/main/java/com/programmers/pcquotation/domain/sellers/dto/SellerUpdateDto.java
@@ -1,4 +1,4 @@
-package com.programmers.pcquotation.sellers.DTO;
+package com.programmers.pcquotation.domain.sellers.dto;
 
 import jakarta.validation.constraints.AssertTrue;
 import jakarta.validation.constraints.NotNull;
