+++ backend/src/main/java/com/programmers/pcquotation/domain/sellers/repository/SellersRepository.java
@@ -1,11 +1,11 @@
-package com.programmers.pcquotation.sellers.repository;
+package com.programmers.pcquotation.domain.sellers.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.programmers.pcquotation.sellers.entitiy.Sellers;
+import com.programmers.pcquotation.domain.sellers.entitiy.Sellers;
 
 @Repository
 public interface SellersRepository extends JpaRepository<Sellers, Long> {
